app/requirements.txt
pytz==2020.5
numpy==1.19.4
python-roku==3.1.5
PlexAPI==4.2.0
paho-mqtt==1.5.1
matplotlib==3.3.3
python-dotenv